pyproject.toml
[project]
name = "gstria-ppg-batch-load"
version = "0.1.0"
description = "Batch load .tbl files into PostgreSQL with partition handling"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "click>=8.1.7",
    "python-dotenv>=1.0.1",
]

[project.scripts]
# 基础命令 -> 指向 main.py 中的 cli_standard 函数
gstria-ppg-batch-load = "gstria_ppg_batch_load.main:cli_standard"
# Collatec 命令 -> 指向 main.py 中的 cli_collatec 函数
gstria-ppg-batch-load-collatec = "gstria_ppg_batch_load.main:cli_collatec"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src/gstria_ppg_batch_load",
    # 注意：删除了 gstria_ppg_batch_load_collatec，因为源码已经合并
]